~ | fix on seq increment, timeval_diff_ms casted as double

+++ srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: guphilip <guphilip@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/24 14:22:56 by guphilip          #+#    #+#             */
-/*   Updated: 2025/11/25 15:31:26 by guphilip         ###   ########.fr       */
+/*   Updated: 2025/11/25 15:35:28 by guphilip         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 /// @return La difference start - end exprime en ms
 double timeval_diff_ms(struct timeval *start, struct timeval *end)
 {
-    return (end->tv_sec - start->tv_sec) * 1000 + 
-            (end->tv_usec - start->tv_usec) / 1000;
+    return (double)(end->tv_sec - start->tv_sec) * 1000 + 
+            (double)(end->tv_usec - start->tv_usec) / 1000;
 }
 
 void print_reply(t_ping *ping, const struct sockaddr_in *from, int bytes, int ttl, double rtt)
